corrected bug admin
