+++ base/tests/models/test_tutor.py
@@ -1,6 +1,6 @@
 ##############################################################################
 #
-#    OSIS stands for Open Student Information System. It's an application
+# OSIS stands for Open Student Information System. It's an application
 #    designed to manage the core business of higher education institutions,
 #    such as universities, faculties, institutes and professional schools.
 #    The core business involves the administration of students, teachers,
@@ -15,7 +15,7 @@
 #
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 #    GNU General Public License for more details.
 #
 #    A copy of this license - GNU General Public License - is available
@@ -23,36 +23,11 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from django.db import models
-from django.contrib import admin
-from django.core.exceptions import ObjectDoesNotExist
-from django.contrib.postgres.fields import JSONField
-from osis_common.models.serializable_model import SerializableModel
+from base.models import tutor
+from base.tests.models import test_person
 
 
-class ScoreEncodingAdmin(admin.ModelAdmin):
-    list_display = ('global_id',)
-    fieldsets = ((None, {'fields': ('global_id', 'document')}),)
-
-
-class ScoreEncoding(SerializableModel):
-    global_id = models.CharField(max_length=10, unique=True)
-    document = JSONField()
-
-    def __str__(self):
-        return u"%s" % self.global_id
-
-
-def find_by_global_id(global_id):
-    try:
-        return ScoreEncoding.objects.get(global_id=global_id)
-    except ObjectDoesNotExist:
-        return None
-
-
-def insert_or_update_document(global_id, document):
-    score_encoding_object, created = ScoreEncoding.objects.update_or_create(
-        global_id=global_id, defaults={"document": document}
-    )
-    return score_encoding_object
-
+def create_tutor():
+    a_tutor = tutor.Tutor(person=test_person.create_person())
+    a_tutor.save()
+    return a_tutor
