+++ dashboard/tests/models/test_score_encoding.py
@@ -25,12 +25,13 @@
 ##############################################################################
 from django.test import TestCase
 from dashboard.models import score_encoding as mdl_score_encoding
-from dashboard.tests import data_for_tests
+import json
+import datetime
 
 
 class ScoreEncodingTest(TestCase):
     def setUp(self):
-        self.score_encoding = data_for_tests.create_score_encoding()
+        self.score_encoding = create_score_encoding()
         self.global_id = self.score_encoding.global_id
 
     def test_find_by_global_id(self):
@@ -40,15 +41,8 @@ class ScoreEncodingTest(TestCase):
         score_encoding = mdl_score_encoding.find_by_global_id(global_id="101245")
         self.assertIsNone(score_encoding, "Should return no score encoding")
 
-    def test_get_document(self):
-        document = mdl_score_encoding.get_document(global_id=self.global_id)
-        self.assertJSONEqual(document, self.score_encoding.document, "Wrong document returned")
-
-        document = mdl_score_encoding.get_document(global_id="120545")
-        self.assertIsNone(document, "Should return no document")
-
     def test_insert_or_update_document(self):
-        new_document = data_for_tests.get_sample()
+        new_document = get_sample()
         score_encoding = mdl_score_encoding.insert_or_update_document("1202151", new_document)
         self.assertJSONEqual(score_encoding.document, new_document, "Problem when inserting new document")
 
@@ -57,8 +51,39 @@ class ScoreEncodingTest(TestCase):
         self.assertJSONEqual(self.score_encoding.document, new_document, "Problem when updating document")
 
 
+sample_1 = "dashboard/tests/ressources/score_encoding_sample.json"
 
 
+def create_score_encoding(global_id="001254"):
+    document = get_sample()
+    score_encoding = mdl_score_encoding.ScoreEncoding(global_id=global_id, document=document)
+    score_encoding.save()
+    return score_encoding
+
+
+def load_sample(sample_path):
+    with open(sample_path) as file_sample:
+        return file_sample.read()
+
+
+def get_sample():
+    sample = get_old_sample()
+    return update_publication_date(sample)
+
+
+def get_old_sample():
+    return load_sample(sample_1)
+
+
+def update_publication_date(sample):
+    json_sample = json.loads(sample)
+    json_sample['publication_date'] = get_today_date()
+    return json.dumps(json_sample)
+
+
+def get_today_date():
+    now = datetime.datetime.now()
+    return '%s/%s/%s' % (now.day, now.month, now.year)
